# Backend/src/main/resources/openapi/device-manager-openapi.yaml
openapi: 3.0.3
info:
  title: Device Manager API
  version: 1.0.0
  description: >
    REST API for managing groups, users, menus, applications,
    wallpapers and themes in the Device Manager system.
  contact:
    name: Milán Buha
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/users/{userId}/menu:
    get:
      tags:
        - Menus
      summary: Get user's root menu
      operationId: getRootMenu
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      responses:
        '200':
          description: Root menu returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RootMenuResponse'
  /api/users/{userId}/menu/applications:
    post:
      tags:
        - Menus
      summary: Add application to root menu
      operationId: addApplicationToRootMenu
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddApplicationRequest'
      responses:
        '201':
          description: Application added successfully
  /api/users/{userId}/menu/applications/{menuItemId}:
    delete:
      tags:
        - Menus
      summary: Remove application from root menu
      operationId: removeApplicationFromRootMenu
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
        - name: menuItemId
          in: path
          required: true
          description: Unique identifier of the menu item
          schema:
            type: string
      responses:
        '204':
          description: Application removed successfully
  /api/users/{userId}/menu/applications/{menuItemId}/launch:
    post:
      tags:
        - Menus
      summary: Launch application from root menu
      operationId: launchApplicationFromRootMenu
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
        - name: menuItemId
          in: path
          required: true
          description: Unique identifier of the menu item
          schema:
            type: string
      responses:
        '200':
          description: Application launched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchResponse'
  /api/users/{userId}/submenus:
    get:
      tags:
        - Menus
      summary: List user's submenus
      operationId: getSubMenus
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      responses:
        '200':
          description: Submenus returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubMenuResponse'

    post:
      tags:
        - Menus
      summary: Create submenu
      operationId: createSubMenu
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubMenuRequest'
      responses:
        '201':
          description: Submenu created successfully
  /api/users/{userId}/submenus/{submenuId}:
    delete:
      tags:
        - Menus
      summary: Delete submenu
      operationId: deleteSubMenu
      parameters:
        - name: userId
          in: path
          required: true
          description: Unique identifier of the user
          schema:
            type: string
        - name: submenuId
          in: path
          required: true
          description: Unique identifier of the submenu
          schema:
            type: string
      responses:
        '204':
          description: Submenu deleted successfully
  /api/submenus/{submenuId}:
    get:
      tags:
        - Menus
      summary: Get submenu details
      operationId: getSubMenu
      parameters:
        - name: submenuId
          in: path
          required: true
          description: Unique identifier of the submenu
          schema:
            type: string
      responses:
        '200':
          description: Submenu returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RootMenuResponse'
  /api/submenus/{submenuId}/applications:
    post:
      tags:
        - Menus
      summary: Add application to submenu
      operationId: addApplicationToSubMenu
      parameters:
        - name: submenuId
          in: path
          required: true
          description: Unique identifier of the submenu
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddApplicationRequest'
      responses:
        '201':
          description: Application added successfully
  /api/submenus/{submenuId}/applications/{menuItemId}:
    delete:
      tags:
        - Menus
      summary: Remove application from submenu
      operationId: removeApplicationFromSubMenu
      parameters:
        - name: submenuId
          in: path
          required: true
          description: Unique identifier of the submenu
          schema:
            type: string
        - name: menuItemId
          in: path
          required: true
          description: Unique identifier of the menu item
          schema:
            type: string
      responses:
        '204':
          description: Application removed successfully
  /api/submenus/{submenuId}/applications/{menuItemId}/launch:
    post:
      tags:
        - Menus
      summary: Launch application from submenu
      operationId: launchSubMenuApplication
      parameters:
        - name: submenuId
          in: path
          required: true
          description: Unique identifier of the submenu
          schema:
            type: string
        - name: menuItemId
          in: path
          required: true
          description: Unique identifier of the menu item
          schema:
            type: string
      responses:
        '200':
          description: Application launched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LaunchResponse'
...
components:
  schemas:
...
    CreateSubMenuRequest:
      type: object
      description: Request body for creating a new submenu
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the submenu
          example: Games
    AddApplicationRequest:
      type: object
      description: Request body for adding an application to a menu or submenu
      required:
        - applicationId
      properties:
        applicationId:
          type: string
          description: Identifier of the application to add
          example: app-openmap
...
    LaunchResponse:
      type: object
      description: Response returned after launching an application
      properties:
        message:
          type: string
          description: Launch result message
          example: "Alkalmazás elindítva: Paint"
    MenuItemResponse:
      type: object
      description: Represents an application item displayed in a menu
      properties:
        id:
          type: string
          description: Unique identifier of the menu item
          example: menu-item-1
        name:
          type: string
          description: Display name of the menu item
          example: Paint
        positionIndex:
          type: integer
          format: int32
          description: Position index of the item inside the menu
          example: 1
    SubMenuResponse:
      type: object
      description: Short summary of a submenu
      properties:
        id:
          type: string
          description: Unique identifier of the submenu
          example: submenu-games
        name:
          type: string
          description: Name of the submenu
          example: Games
    RootMenuResponse:
      type: object
      description: Detailed representation of a user's root menu or submenu
      properties:
        menuName:
          type: string
          description: Name of the menu
          example: Buha főmenüje
        wallpaperName:
          type: string
          nullable: true
          description: Currently selected wallpaper name
          example: Mountain
        themeName:
          type: string
          nullable: true
          description: Currently selected theme name
          example: Dark
        applications:
          type: array
          description: Applications displayed directly in the menu
          items:
            $ref: '#/components/schemas/MenuItemResponse'
        subMenus:
          type: array
          description: Submenus accessible from the root menu
          items:
            $ref: '#/components/schemas/SubMenuResponse'
...